requirements.txt
Flask==2.2.2
pandas==1.5.3
numpy==1.24.2
scikit-surprise==1.1.1
tensorflow==2.12.0